Remove duplicated struct
