+++ Views/ViewModifiers/SaveContentHandler.swift
@@ -93,28 +93,6 @@ struct SaveContentHandler: ViewModifier {
     #endif
 }
 
-#if os(iOS)
-struct ActivityViewController: UIViewControllerRepresentable {
-
-    var activityItems: [Any]
-    @Environment(\.dismiss) var dismissAction
-    func makeUIViewController(
-        context: UIViewControllerRepresentableContext<ActivityViewController>
-    ) -> UIActivityViewController {
-        let controller = UIActivityViewController(activityItems: activityItems, applicationActivities: nil)
-        controller.modalPresentationStyle = .pageSheet
-        controller.completionWithItemsHandler = { (_, _, _, _) in
-            self.dismissAction()
-        }
-        return controller
-    }
-
-    func updateUIViewController(_ uiViewController: UIActivityViewController, context: Context) {
-    }
-}
-
-#endif
-
 extension URL {
     fileprivate func tempFileURL() -> URL? {
         let directory = FileManager.default.temporaryDirectory
